+++ pyKVFinder/grid/__init__.py
@@ -1,0 +1,32 @@
+from .cavity import _get_cavity_label, _get_cavity_name, _select_cavities
+from .constitutional import _process_residues, constitutional
+from .depth import _process_depth, depth
+from .detect import detect
+from .export import export, export_openings
+from .geometry import (_get_dimensions, _get_sincos, _get_vertices_from_box,
+                       _get_vertices_from_residues, get_vertices,
+                       get_vertices_from_file)
+from .hydropathy import _process_hydropathy, hydropathy
+from .openings import (_get_opening_label, _get_opening_name,
+                       _process_openings, openings)
+from .spatial import _process_spatial, spatial
+
+# Public API: only non-private functions are exposed for backward compatibility with v0.8.x
+
+__all__ = [
+    "_get_cavity_label",
+    "_get_cavity_name",
+    "_select_cavities",
+    "constitutional",
+    "depth",
+    "detect",
+    "export",
+    "export_openings",
+    "_get_dimensions",
+    "_get_sincos",
+    "get_vertices",
+    "get_vertices_from_file",
+    "hydropathy",
+    "openings",
+    "spatial",
+]
