+++ pyproject.toml
@@ -88,8 +88,11 @@ issues = "https://github.com/LBC-LNBio/pyKVFinder/issues"
 pyKVFinder = "pyKVFinder.main:cli"
 
 [tool.setuptools]
-packages = ["pyKVFinder", "pyKVFinder.data", "pyKVFinder.data.tests"]
+packages = ["pyKVFinder", "pyKVFinder.grid"]
 include-package-data = true
+
+[tool.setuptools.package-data]
+pyKVFinder = ["data/*.toml", "data/vdw.dat", "data/tests/*"]
 
 [tool.setuptools.dynamic]
 version = { attr = "pyKVFinder.__version__" }
